+++ src/main/java/com/kernel360/boogle/book/controller/BookController.java
@@ -3,10 +3,7 @@ package com.kernel360.boogle.book.controller;
 import com.kernel360.boogle.book.db.BookEntity;
 import com.kernel360.boogle.book.model.BookDTO;
 import com.kernel360.boogle.book.service.BookService;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
 import java.util.ArrayList;
@@ -42,5 +39,13 @@ public class BookController {
     @GetMapping("/admin/book")
     public List<BookEntity> getBook() {
         return bookService.findAllBook();
+    }
+
+    @PatchMapping("/admin/book")
+    public void updateBook(
+            @RequestBody
+            BookDTO book
+    ){
+        bookService.updateBook(book);
     }
 }
