+++ src/main/java/com/kernel360/boogle/book/model/BookDTO.java
@@ -13,4 +13,6 @@ import lombok.NoArgsConstructor;
 public class BookDTO {
 
     private BookEntity bookEntity;
+
+
 }
