+++ src/main/java/com/kernel360/boogle/book/service/BookService.java
@@ -1,26 +1,15 @@
 package com.kernel360.boogle.book.service;
 
-import com.kernel360.boogle.book.model.BookDTO;
 import com.kernel360.boogle.book.db.BookEntity;
-import com.kernel360.boogle.book.db.BookRepository;
-import org.springframework.stereotype.Service;
-import javax.transaction.Transactional;
-import lombok.AllArgsConstructor;
+import com.kernel360.boogle.book.model.BookDTO;
 
 import java.util.List;
 
-@Service
-@Transactional
-@AllArgsConstructor
-public class BookService {
-    private final BookRepository bookRepository;
+public interface BookService {
+    public List<BookEntity> findAllBook();
 
-    public List<BookEntity> findAllBook() {
-        return bookRepository.findAll();
-    }
+    public void saveBook(BookDTO book);
 
-    public void saveBook(BookDTO book) {
-        bookRepository.save(book.getBookEntity());
-    }
+    public void updateBook(BookDTO book);
 
 }
