+++ src/main/resources/application.yml
@@ -8,6 +8,6 @@ spring:
   #      ddl-auto: validate
   datasource:
     driver-class name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/boogle?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true&createDatabaseIfNotExist=true
-    username: root
-    password: ""
+    url: jdbc:mysql://e2e-boogle-db.ck74at0lmilr.us-east-1.rds.amazonaws.com:3306/E2E_BOOGLE?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true&createDatabaseIfNotExist=true
+    username: admin
+    password: admin123$
